requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
langchain
PyPDF2
python-docx
google-generativeai
langchain-community
langchain-google-genai
faiss-cpu
python-dotenv
